CRITICAL Bugfix for display list or immediate mode rendering
If transfer mode was display list or immediate mode, PRREObject3D::PRREObject3DImpl::Draw() returned before resetting render mode from feedback to render. This caused the object not visible on screen, since it was always being rendered in feedback mode. This happened only if this was the only object in the scene with display list or immediate mode, since rendering a different object in a different mode would had reset rendering mode to render from feedback.