PRRETexture creating from file fix memory corruption and unneccessary memory usage
There was a memory leak and possible memory corruption opportunity with PRRETexture loading from file. When loading from file, first a PRREImage was created with the pixel data loaded from the file, and from that a PRRETexture was created. But unfortunately, the PRRETexture having its own PRREImage part, got it updated from the original PRREImage, meaning that it stored the same pointers, etc as the same Image, without array copy, so both the original Image and the Texture's Image part pointed to same pixel array. This means, techinally 2 Image instances got created during a Texture loading, and the 1st could be deleted having effect and memory corruption on the 2nd Image instance (part of the Texture instance itself).
For better performance and less confusion, PRREImage::Cannibalize() got introduced, the same method we used in PRREMesh.