UnitTest::setUp() must return result
UnitTest::setUp() now has return type bool, it should be true if setUp() process succeeded. If it returns false, TestMethod() and subtests will be skipped.
This way there is no need to check for resources availability in each tests, that supposed to be loaded by setUp(), since setUp()'s responsibility is to return proper value based on if it loaded resources or not.