Slighly modified Mesh3D vertex index related functions
getMinIndexValue() renamed to getMinVertexIndex();
getMaxIndexValue() renamed to getMaxVertexIndex();
SetIndexInArray() renamed to SetVertexIndex(), arr parameter is removed since it is now straightforward;
getIndexFromArray renamed to getVertexIndex(), arr parameter is removed since it is now straightforward, and now it checks if the given index is in the valid range, also has the implicitAccessSubobject parameter as well with default true.
Updated the related unit tests as well.