Documentation update
There is a bug in doxygen about the relative path of images that should be moved to the output html directory. Originally I used an img_common directory for images shared between external and internal documentation. That dir was in docpages directory, so the included images in docpages/internal/*.md were included with path ../img_common/ . These were NOT copied to output html directory by doxygen! Only the internal images stored in docpages/internal/img/, for which I used path img/ in .md files.
There is an issue opened and already a pull request to fix this issue:
https://github.com/doxygen/doxygen/pull/6787
But for some reason it is still not part of the official build (1.8.20). So as a workaround, I moved docpages/img_common directory to docpages/internal/, so both the img_common and img dirs are in same directory. This img_common directory needs to be duplicated into docpages/external/ dir, this is the drawback of this workaround, but hopefully soon the doxygen fix will be part of the official latest build!