Preparation for classes invoking CConsole with their unique name
In order to introduce per-module logging filter, CConsole must be aware of the current class doing the logging. So I've added a string param to CConsole::getConsoleInstance(), which sets the current logging module name inside CConsole, so CConsole knows who invoked it.
I've changed the getConsole() function of all classes to use this parameter when they invoke CConsole::getConsoleInstance().
So now CConsole is always aware of WHO invoked it.
Filtering is not yet implemented, but in a next commit, per-module filtering will be available with this approach: if the current logging invoker module is enabled for logging, then CConsole will log, otherwise it will do nothing.

I've also deleted PRREBaseClass from everywhere. This is why PRREManaged.h now includes PRREtypes.h.

Also, in a next commit, PRREManaged::getManagedConsole() will need to be renamed to getConsole(). This will introduce ambiguity in classes which are derived from both Managed and Manager, so virtual inheritance will be also needed!