Add EV2008P03 Sample
EV2008P03 is task 03 of EngineVerseny 2008. This is included as a sample, although it needs gfxcore2 adapter layer to use PRRE. We are going to use this application as periodical performance measurement of different versions of the engine.

Adding new project configurations EV2008P03Debug and EV2008P03Release, so it is easy to build gfxcore2 and run EV2008P03 with it with 1 click.