Add BVH class and misc stuff
Added PRREBoundingVolumeHierarchy class. Note that it is still not complete, its insertObject() doesn't update AABBs, and the unit test is just copy paste from PRREOctreeTest, so I return false there intentionally to mark it as faulty.

Made default ctor of PRREAxisAlignedBoundingBox public, which means that it can easily created with size (0,0,0), but this needs to be unit tested !!!

Added parent member to PRREOctree, so children know their parent. This will be useful in derived PRREBoundingVolumeHierarchy's insertObject() that will be able to refresh the AABB of the parent nodes all the way up to root of that node that stored the inserted object. This way we don't need a separate build() function that must be called after all insertObject()s to make the BVH actually functional. And this is still fast. Without parent ptr, only slower implementation would be possible.