Simple Sync Occlusion Query Implementation
In contrast to a previous sync occlusion query commit happened this week, in this commit, the rendering is happening in 2 passes:
 - 1st pass: only bounding boxes of complex objects are rendered, without writing anything to z- or color buffer, queries are started with this;
 - 2nd pass: render simple objects, and if occlusion query of complex object pass, render that too.

The big advantage of this commit is that we turn off writing to color and depth buffers only once, this is the 1st pass. During the start of queries, there is chance that some previous queries are finished in the meantime.
Then in 2nd pass we enable depth and color write. In previous commit this disable/enable stuff happened per every object! Then we wait for query result of object and might draw the object. If we draw the object, other pending queries might be also finished in the meantime.

Overall, this commit stalls the pipeline less much compared to the other commit.
This approach stays as the simple sync occlusion query implementation. The next technique must be an async approach!