Fixes around pBoundingBox
Fix 1: the rel pos vec was forgot to be applied in case of createCloned()!

Fix 2: in case of cloned object, its getRelPosVec() should return the rel pos vec of the referred object.

Fix 3: the box created for pBoundingBox should use DYNAMIC VMOD, otherwise the forceUseClientMemory = true parameter will be ignored.