Attach(), Detach() override to handle occlusion stuff
When an Object3D is detached from its manager, it should be removed from both occludee and occluder list, also its isoccluder and isocclusiontested states should be set to false.

When an Object3D is re-attached to Object3DManager, it should be added to occludee list.

Object3D's SetOcclusionTested() and SetOccluder() must reject when it has no manager associated.