FTR: Finally using the separate containers in PRRERendererHWfixedPipeImpl
Also some fixes:

Fix 1:
2D objects were also put previously into either occludee_... lists, so they were accidentally drawn in 3D mode as well due to this commit, so I rather introduced separate containers for 2D objects.

Fix 2:
due to blending mode for all objects was set with getMaterial() call, for 2D objects this meant implicit subobject access, so their 0th level-2 subobject material was changed, but this is a problem because then it did not invoke Object3DManager's HandleManagedPropertyChanged() function. So everywhere I changed getMaterial() to getMaterial(false) to make sure we always set the blending mode for the level 1 object.

Fix 3:
Due to the abovementioned fix 2,PRREGLsnippets:: glLoadTexturesAndSetBlendState() accidentally turned blending off in some cases when an object was NOT multitextured. This is because this function is invoked only for subobjects, and after fix 2 the subobjects no longer had their material holding blending property, so it should be also aware of level-1 object's material blending property.

Next steps left out from this commit:
 - UpdateOccluderStates() still has getMaterial() without parameter!
 - getMaterial() calls without parameter must be explicitly changed to getMaterial(false) when accessing properties other than blending, like colorkey stuff. I'm also thinking that on the long run maybe we should get rid of this argument.
 - add functions for accessing the 2 new containers: get2dOpaqueOccludees() and get2dBlendedOccludees().