Various fixes around occlusion culling and engine debug switches
Fix 1: implement ForceFinishOcclusionTest(). SetRenderHints() invokes this for all objects when rendering path or occlusion query mode is changed during runtime, to make sure all pending queries are finished and we are not inconsistent when user triggers a debug switch.

Fix 2: SetOcclusionTested() should not create bounding box object when we are a cloned object, and referred object already has bounding box. During rendering, additional checks are added to use referred object's bounding box in such case. This change speeds up cloning occlusion tested objects.

Fix 3: fix the nFramesWithoutOcclusionTest related checks in Draw_OcclusionQuery_Start() so occlusion queries are now properly started in that frame when they need to, not in the next frame! Added 2 asserts in the code to abort the program in debug mode in case of programmer failure.