Add PRREObject3D::getReferrerObjects(), enhance SetOcclusionTested()
The referred object should also have a list of referrer objects, so when a referred object is destructed, it sets the referring ptr to NULL in all of the referred objects so they won't hold invalid ptr anymore. The referring objects are NOT deleted in this case though - it is the user's responsibility to delete them too. This change makes the engine more robust though by avoiding crash due to invalid ptr.

The list of referrers can be get by getReferrerObjects().

SetOcclusionTested() cannot turn occlusion testing off for an object which has referrer objects - this is because those referrer objects might still have occlusion testing on and in such case they might rely on the bounding box object of the original referred object. That bounding box would be otherwised deleted without this check. Note that, on the long run, we might iterate over the referrers and turn their occlusion testing off too - but this has not been decided yet.