Fix bounding box related memory leaks detected at engine shutdown
SetOcclusionTested(false) now implicitly turns occlusion testing off for all referring cloned objects too, so the bounding box of the original object can be safely deleted. This gets rid of the memory leak where the dtor of original object invokes SetOcclusionTested(false) but that left the bounding box in memory with previous behavior.