Rationalize Mesh3D::getVerticesCount()
Remove the implicitAccessSubobject argument, it is confusing and doesn't give easier coding anyway now since we introduce the following behavior:
if this function is invoked for level-1 object, it returns the sum of number of vertices of all its level-2 subobjects.

Fixed some of the getUsedVideoMemory() and getUsedSystemMemory() functions.

Added some asserts to some functions.