FIX: TextureManager's LazyInstancing property was behaving inverted
Value of bAlwaysCreateNewFromFile was false by default, however isLazyInstancingEnabled() returned the variable as is instead of returning the negated value of the variable. Hence for the same filename and path always a new texture was created. Leading to ~120 MiByte texture memory usage in EV2008P03 instead of ~32 MiByte. I fixed this, including its unit tests ...