FIX: Blending for Cloned Objects
Blending for cloned objects was not working if the referred object itself was not blended.

This was because when a cloned object was drawn, although it set blending and other config based on the cloned object's config, however before the referred object started to iterate over its subobjects, it invokes additional logic (PureGLsnippets:: glLoadTexturesAndSetBlendState) which takes the subobject's parent (i.e. the referred object) blending config into account instead of the referee's blending config.

Solution is to always use the referee object's blending config even when referred object is iterating over its subobjects for transfering their vertices.